backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
fastapi==0.104.1
greenlet==3.1.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.10
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.1
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.13.1
uvicorn==0.23.2
